requirements.txt
# fasthtml
python-fasthtml
monsterui
fasthtml-hf
